requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Time series forecasting
prophet>=1.1.0
scikit-learn>=1.3.0
statsmodels>=0.13.0

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dateutil>=2.8.0
plotly
nbformat
holidays
ruff
PyYAML>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0
